requirements.txt
-c requirements.lock


Django
djangorestframework
django-trix-fork
